fix: PR Comment 반영 폴더 구조 변경 및 삭제 Related: #41
