+++ src/main/java/com/example/dgbackend/domain/combination/Combination.java
@@ -1,23 +1,24 @@
 package com.example.dgbackend.domain.combination;
 
 
+import com.example.dgbackend.domain.combinationcomment.domain.CombinationComment;
+import com.example.dgbackend.domain.combinationimage.CombinationImage;
 import com.example.dgbackend.domain.member.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.DynamicInsert;
+import org.hibernate.annotations.DynamicUpdate;
 
+import java.util.ArrayList;
+import java.util.List;
+
+@Builder
 @Getter
+@DynamicInsert
+@DynamicUpdate
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
@@ -39,10 +40,37 @@ public class Combination extends BaseTimeEntity {
     @ColumnDefault("0")
     private Long commentCount;
 
+    @Builder.Default
     private boolean state = true; //true : 존재, false : 삭제
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
+    @OneToMany(mappedBy = "combination", cascade = CascadeType.ALL)
+    private List<CombinationImage> combinationImages = new ArrayList<>();
+
+    @OneToMany(mappedBy = "combination", cascade = CascadeType.ALL)
+    private List<CombinationComment> combinationComments = new ArrayList<>();
+
+    /**
+     * 연관관계 편의 메소드
+     */
+    public void addCombinationImage(CombinationImage combinationImage) {
+        combinationImages.add(combinationImage);
+        combinationImage.setCombination(this);
+    }
+
+    public void addCombinationComment(CombinationComment combinationComment) {
+        combinationComments.add(combinationComment);
+        combinationComment.setCombination(this);
+    }
+
+    /**
+     * update 함수
+     */
+    public void updateCombination(String title, String content) {
+        this.title = title;
+        this.content = content;
+    }
 }
