+++ src/main/java/com/example/dgbackend/domain/combination/repository/CombinationRepository.java
@@ -1,0 +1,7 @@
+package com.example.dgbackend.domain.combination.repository;
+
+import com.example.dgbackend.domain.combination.Combination;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CombinationRepository extends JpaRepository<Combination, Long> {
+}
