+++ src/main/java/com/example/dgbackend/domain/combinationimage/repository/CombinationImageRepository.java
@@ -1,6 +1,6 @@
 package com.example.dgbackend.domain.combinationimage.repository;
 
-import com.example.dgbackend.domain.combinationimage.domain.CombinationImage;
+import com.example.dgbackend.domain.combinationimage.CombinationImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
