+++ src/main/java/com/example/dgbackend/domain/combinationimage/service/CombinationImageCommandService.java
@@ -1,6 +1,6 @@
 package com.example.dgbackend.domain.combinationimage.service;
 
-import com.example.dgbackend.domain.combination.domain.Combination;
+import com.example.dgbackend.domain.combination.Combination;
 
 import java.util.List;
