+++ src/main/java/com/example/dgbackend/domain/combinationlike/CombinationLike.java
@@ -1,7 +1,7 @@
 package com.example.dgbackend.domain.combinationlike.domain;
 
-import com.example.dgbackend.domain.combination.domain.Combination;
-import com.example.dgbackend.domain.member.domain.Member;
+import com.example.dgbackend.domain.combination.Combination;
+import com.example.dgbackend.domain.member.Member;
 import com.example.dgbackend.global.common.BaseTimeEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
