+++ src/main/java/com/example/dgbackend/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.example.dgbackend.domain.member.repository;
 
-import com.example.dgbackend.domain.member.domain.Member;
+import com.example.dgbackend.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
